pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "pardoc"
version = "0.0.1"
description = "Yet another docstring parser for python"
authors = [ "pwwang <pwwang@pwwang.com>",]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pwwang/pardoc"
repository = "https://github.com/pwwang/pardoc"

[tool.poetry.dependencies]
python = "^3.6"
lark-parser = "^0.8"
diot = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
